_project/aistudio/requirements.txt
streamlit
twikit
crewai
crewai-tools
langchain-groq
python-dotenv
